Add solution for problem `F - Anton and School` Link: https://codeforces.com/contest/734/problem/F Tags: bitmasks, constructive algorithms, implementation, math, *2500 Ref: https://codeforces.com/contest/734/submission/22296942

+++ codeforces/734/F.cpp
@@ -4,7 +4,7 @@ using namespace std;
 typedef long long ll;
 
 const int N = 200005;
-int a[N];
+ll a[N];
 int b[N];
 int c[N];
